Revert dtype to float16 in 7B baseline script

+++ experiments/wk02_7b_baseline/a100_7b.py
@@ -15,7 +15,7 @@ def run_7b():
     model_id = "Qwen/Qwen2.5-7B-Instruct"
     tokenizer = AutoTokenizer.from_pretrained(model_id)
     model = AutoModelForCausalLM.from_pretrained(
-        model_id, dtype=torch.float32
+        model_id, dtype=torch.float16
     ).cuda()
 
     prompt = "The key to efficient LLM inference is"
